requirements.txt
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
numpy==1.26.4
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
Telethon==1.34.0
urllib3==2.2.0
